Solution to Check Greater Than

+++ Check Greater Than@@67882a731b4392d8fc1577d6/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    int m,n;
-    scanf("%d",&m, &n);
+    float m,n;
+    scanf("%f",&m, &n);
     if(m>n){
         printf("True");
     }
